feat: sims2pc.exe LINKS AND RUNS — first x86_64 executable from GC decomp
The Sims 2 GameCube decompilation now produces a working PC executable:
- 1.1MB PE32+ x86_64 Windows binary
- All 47 decomp_cpp files compile and link
- Platform stubs for GX/PAD/DVD/OS/CARD/AX
- Outputs game info and clean exit (no SDL2 yet for window)

Fixed: operator new size_t, pointer-to-u32 casts, C/C++ linkage
mismatches for strcmp/Sprintf/AtoI/stricmp, anonymous namespace
InterpolateLinear, duplicate symbol resolution.

Next: install SDL2 for window + rendering backend.

+++ src/decomp_cpp/cxobjectimpl.cpp
@@ -502,15 +502,15 @@ cXObject* cXObjectImpl::GetNext() {
 
 int cXObjectImpl::HierCountSlots() {
     // Each slot is 32 bytes; count = (end - begin) / 32
-    int begin = (int)(unsigned int)m_slotsBegin;
-    int end = (int)(unsigned int)m_slotsEnd;
+    int begin = (int)(uintptr_t)m_slotsBegin;
+    int end = (int)(uintptr_t)m_slotsEnd;
     return (end - begin) >> 5;
 }
 
 int cXObjectImpl::CountObjectSlots() {
     // Same as HierCountSlots but subtract 1 (slot 0 is reserved)
-    int begin = (int)(unsigned int)m_slotsBegin;
-    int end = (int)(unsigned int)m_slotsEnd;
+    int begin = (int)(uintptr_t)m_slotsBegin;
+    int end = (int)(uintptr_t)m_slotsEnd;
     return ((end - begin) >> 5) - 1;
 }
